Change host_path result in broken test to null due to algorithm change
The path does not exist. In this case, with the new algorithm returned
path will be empty. It makes sense since there is no active path.
Also the old algorithm does not gurantee to return the real path the
packet is taking. Even if it printed half path, it is not ensured that
the path is the actual path.